Trying to improve code quality
* Code currently broken :-(

+++ src/main/java/DeleteCommand.java
@@ -1,29 +1,58 @@
 import java.io.IOException;
 
 public class DeleteCommand extends Command {
+
     protected int deleteIdx;
     public DeleteCommand(int deleteTaskIdx) {
         super("delete");
         this.deleteIdx = deleteTaskIdx;
     }
 
-    public void execute(TaskList taskList, Ui ui, Storage storage) throws IOException {
+    /**
+     * Deletes task from list, saves changes in text file.
+     * @param taskList
+     * @param ui
+     * @param storage
+     * @throws DukeException
+     * @throws IOException
+     */
+    public void execute(TaskList taskList, Ui ui, Storage storage) throws DukeException, IOException {
+        // Checks if index of task to delete is within range
+        int numTasks = taskList.getNumTasks();
+        if (numTasks == 0) {
+            throw new DukeException("You have no tasks to delete!");
+        }
+        if (deleteIdx < 1 || deleteIdx > numTasks) {
+            throw new DukeException("Please provide a positive integer that is " +
+                    numTasks + " and below.");
+        }
+
         Task deletedTask = taskList.deleteTask(deleteIdx);
         storage.save(taskList.getTaskArr());
-        ui.showDeleteTaskMessage(deletedTask, taskList.getTaskArr());
+        ui.showDeleteTaskResponse(deletedTask, taskList.getTaskArr());
+    }
+
+    public int getDeleteIdx() {
+        return deleteIdx;
     }
 
     public void print() {
         super.print();
-        System.out.println("Index of task to delete: " + deleteIdx);
+        System.out.println("Index of task to mark done: " + deleteIdx);
     }
 
     @Override
     public boolean equals(Object o) {
-        if (o == this) { return true; }
-        // Not even the same class
-        if (!(o instanceof DeleteCommand)) { return false; }
+        // Not even same class
+        if (!(o instanceof Command)) {
+            return false;
+        }
+        // Same object
+        if (o == this) {
+            return true;
+        }
         DeleteCommand c = (DeleteCommand) o;
-        return c.deleteIdx == deleteIdx;
+        boolean sameDeleteIdx = deleteIdx == c.getDeleteIdx();
+        return super.equals(o) && !sameDeleteIdx;
     }
 }
